Small rework

+++ Assets/Project/Scripts/Features/Skin/Shop/SkinShopItemData.cs
@@ -1,11 +1,11 @@
 public class SkinShopItemData
 {
     private readonly int _id;
-    private readonly float _price;
+    private readonly int _price;
     private readonly bool _isUnlocked;
     private readonly bool _canBuy;
 
-    public SkinShopItemData(int id, float price, bool isUnlocked)
+    public SkinShopItemData(int id, int price, bool isUnlocked)
     {
         _id = id;
         _price = price;
@@ -13,6 +13,6 @@ public class SkinShopItemData
     }
 
     public int Id => _id;
-    public float Price => _price;
+    public int Price => _price;
     public bool IsUnlocked => _isUnlocked;
 }
